refactor: query용 service 클래스 분리 (#15)
