+++ src/main/java/com/vestie/vestie/survey/presentation/SurveyController.java
@@ -1,5 +1,6 @@
 package com.vestie.vestie.survey.presentation;
 
+import com.vestie.vestie.survey.application.SurveyQueryService;
 import com.vestie.vestie.survey.application.SurveyService;
 import com.vestie.vestie.survey.presentation.dto.SurveyInquiryResponse;
 import com.vestie.vestie.survey.presentation.dto.SurveyRegisterRequest;
@@ -18,6 +19,7 @@ import static org.springframework.http.HttpStatus.CREATED;
 public class SurveyController {
 
     private final SurveyService surveyService;
+    private final SurveyQueryService surveyQueryService;
 
     @PostMapping
     ResponseEntity<Long> register(@RequestBody SurveyRegisterRequest request) {
@@ -28,13 +30,13 @@ public class SurveyController {
 
     @GetMapping
     ResponseEntity<List<SurveyInquiryResponse>> getAllSurvey() {
-        var allSurvey = surveyService.getAllSurvey();
+        var allSurvey = surveyQueryService.getAllSurvey();
         return ResponseEntity.ok(allSurvey);
     }
 
     @GetMapping("/{surveyId}")
     ResponseEntity<SurveyResponse> getSurvey(@PathVariable Long surveyId) {
-        var survey = surveyService.getSurvey(surveyId);
+        var survey = surveyQueryService.getSurvey(surveyId);
         return ResponseEntity.ok(survey);
     }
 }
